requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Model Export & Inference
onnx>=1.14.0
onnxruntime>=1.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Optional: For GPU support
# onnxruntime-gpu>=1.15.0

# Optional: For Jetson deployment
# jetson-stats>=4.2.0
# pycuda>=2022.1